refactor: modularize the api
