+++ backend/.env.example
@@ -5,6 +5,8 @@ DATABASE_HOST=db
 DATABASE_PORT=5432
 DATABASE_URL=postgresql://${DATABASE_USER}:${DATABASE_PASSWORD}@${DATABASE_HOST}:${DATABASE_PORT}/${DATABASE_NAME}?connect_timeout=1000
 DYNAMODB_URL=http://dynamodb:8000
-AWS_KEY=http://dynamodb:8000
-AWS_REGION=http://dynamodb:8000
-AWS_SECRET=http://dynamodb:8000
+AWS_KEY=AKID
+AWS_REGION=us-east-1
+AWS_SECRET=SECRET
+REDIS_PORT=6379
+REDIS_HOST=redis
