+++ backend/src/datastore/index.ts
@@ -1,0 +1,5 @@
+import { connectDynamodb } from './dynamo'
+import { prisma } from './prisma'
+import { redis } from './redis'
+
+export { connectDynamodb, prisma, redis }
