+++ backend/src/datastore/prisma.ts
@@ -1,0 +1,8 @@
+import {
+  PrismaClient,
+  Prisma,
+  FamilyMember,
+  PolicyStatus,
+} from '@prisma/client'
+
+export const prisma = new PrismaClient()
