+++ backend/src/datastore/redis.ts
@@ -1,0 +1,6 @@
+const Redis = require('ioredis')
+
+export const redis = new Redis({
+  host: process.env.REDIS_HOST || 'redis',
+  port: process.env.REDIS_PORT || 6379,
+})
