+++ backend/src/modules/index.ts
@@ -1,0 +1,6 @@
+import policy from './policy/policy.routes'
+import { Express } from 'express'
+
+export default async (app: Express) => {
+  app.use('/policies', policy)
+}
